requirements.txt
discord
aiohttp
pillow
numpy
easyocr
python-dotenv
pyttsx3==2.90
